chore: rename carousel component
